utility_server/requirements.txt
fastmcp>=0.1.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
asyncio>=3.4.3
